[C] Consolidate qacompound components into one
use QACompoundInterface as core compount parent component.
use questionType instead of answerAccessor.
rename QACompoundInterface to QACompound.
set new QACompound as the compound comp for compoundSelect and compoundInput.
