+++ src/components/qas/QA.jsx
@@ -19,10 +19,10 @@ const questions = {
   text: QATextInput,
   textArea: QATextInput,
   select: QASelect,
-  compound: QACompound,
-  compoundSelect: QASelect,
   'inline-select': QASelect,
   multiSelect: QAMultiSelect,
+  compoundSelect: QACompound,
+  compoundInput: QACompound,
   DistanceCalculator,
   SizeCalculator,
   KineticEnergyCalculator,
