+++ src/components/qas/questions/qaCompound/index.jsx
@@ -1,30 +1,72 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import QACompoundSelect from '../qaCompoundSelect/index.jsx';
-import QACompoundTextInput from '../qaCompoundTextInput/index.jsx';
-import QACompoundInterface from '../qaCompoundInterface/index.jsx';
+import classnames from 'classnames';
+import Card from '../../../site/card';
+import QASelect from '../qaSelect';
+import QATextInput from '../qaTextInput';
+import { qaCard, active } from '../../styles.module.scss';
 
 class QACompound extends React.PureComponent {
   constructor(props) {
     super(props);
-    this.compoundTypes = {
-      compoundInput: QACompoundTextInput,
-      compoundSelect: QACompoundSelect,
-      compound: QACompoundInterface,
+
+    this.qaTypes = {
+      compoundInput: QATextInput,
+      compoundSelect: QASelect,
+    };
+
+    this.state = {
+      hasFocus: false,
     };
   }
 
-  render() {
-    const { questions } = this.props;
-    const { questionType } = questions[0];
-    const CompoundComponent = this.compoundTypes[questionType];
+  updateActive = hasFocus => {
+    this.setState(prevState => ({
+      ...prevState,
+      hasFocus,
+    }));
+  };
 
-    return CompoundComponent ? <CompoundComponent {...this.props} /> : null;
+  render() {
+    const { questions, activeId, answers, handleAnswerSelect } = this.props;
+    const { hasFocus } = this.state;
+    const cardClasses = classnames(qaCard, { [active]: hasFocus });
+
+    return (
+      <Card className={cardClasses}>
+        <div className="qa-wrapper">
+          {questions.map(question => {
+            const { id, questionType, options } = question;
+            const QACompoundComponent = this.qaTypes[questionType];
+
+            return (
+              <QACompoundComponent
+                key={id}
+                ids={question.compoundQuestion}
+                answer={answers[id]}
+                focusCallback={this.updateActive}
+                answerHandler={handleAnswerSelect}
+                handleAnswerSelect={handleAnswerSelect}
+                {...{
+                  question,
+                  activeId,
+                  questionType,
+                  options,
+                }}
+              />
+            );
+          })}
+        </div>
+      </Card>
+    );
   }
 }
 
 QACompound.propTypes = {
+  handleAnswerSelect: PropTypes.func,
   questions: PropTypes.array,
+  answers: PropTypes.object,
+  activeId: PropTypes.string,
 };
 
 export default QACompound;
